IGNITE-24098 Remove PROTO COL_V2_VERSION_SINCE
